+++ HISTORY.rst
@@ -12,6 +12,11 @@ History
 
 **Other Changes:**
 
+* Changed the vcf file generator to emit reference bases in uppercase.  Added the ``vcfPreserveRefCase``
+  flag to the call_consensus.py script to cause the vcf file generator to emit each reference base in
+  uppercase/lowercase as it appears in the original reference sequence file.  If not specified, the
+  reference bases are emitted in uppercase.  Prior to v0.4.2, the behavior was to always preserve the
+  original case.
 * Removed setuptools from the list of required installation dependencies.
 * Added support for Python 3.3, 3.4, 3.5.